[Client] Extract snackbar to hook
